cafeteria_turnos/requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
psycopg2-binary  # Solo si usas PostgreSQL, puedes quitarlo para SQLite
pyjwt